tidy up contribution guide
