+++ internal/docs/sidebar/list.js
@@ -31,7 +31,6 @@ const guides = [
   { path: '/', title: 'Getting started' },
   { path: '/usage', title: 'How to use Cosmos?' },
   { path: '/guiding-principles', title: 'Guiding Principles' },
-  { path: '/system-guidelines', title: 'System Guidelines' },
   { path: '/faqs', title: 'FAQs' },
   { path: '/changes', title: 'Changelog' },
   { path: '/automation', title: 'Automation Glossary' },
